requirements.txt
pywinauto
flask
google-generativeai
